# packages/api-proxy/swagger/swagger-spec-v3-klaytn.yaml
openapi: 3.0.0
info:
  title: KLAYTN
  description: Henesis API Proxy - v3 - klaytn OpenAPI Documentation
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  schemas:
    InvalidAccessIpException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: ip '106.101.128.133' is not allowed
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4008
      required:
        - message
        - code
    InvalidAccessTokenException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: expired or invalid access token and signature
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4001
      required:
        - message
        - code
    AccessTokenNotProvidedException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: access token not provided
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4001
      required:
        - message
        - code
...
    InvalidStatusException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: '''status'' is invalid: not supported event status'
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4001
      required:
        - message
        - code
...
    PaginationMetadata:
      type: object
      properties:
        nextUrl:
          type: string
          description: 다음 pagination url
          example: http://localhost:3000/api/v3/ethereum/transfers?page=2
        previousUrl:
          type: string
          description: 이전 pagination url
          example: http://localhost:3000/api/v3/ethereum/transfers?page=1
        totalCount:
          type: number
          description: 총 갯수
          example: 5
      required:
        - nextUrl
        - previousUrl
        - totalCount
...
    TransferDTO:
      type: object
      properties:
        id:
          type: number
          description: 코인/토큰 입출금 내역의 ID
          example: 375031
        from:
          type: string
          description: 출금 주소
          example: '0x1f10ecbd971eab345ea19e96dc237b1fbd63de96'
        to:
          type: string
          description: 입금 주소
          example: '0xdd00383e5a51166bba3e96d84c14a36d72e6c128'
        amount:
          type: string
          description: 암호화폐의 양
          example: '1000000000'
        status:
          enum:
            - PENDING_APPROVAL
            - REJECTED
            - REQUESTED
            - PENDING
            - FAILED
            - REVERTED
            - REPLACED
            - MINED
            - CONFIRMED
          type: string
          description: 트랜잭션 상태
          example: CONFIRMED
        orgId:
          type: string
          description: 지갑이 속한 팀(Organization)의 ID
          example: 31cafc79bd437e1b8e7b8209e399d3f3
        decimals:
          type: number
          description: 암호화폐의 소수점 자릿수
          example: 18
        walletId:
          type: string
          description: 입출금 지갑 ID
          example: ae40b1b3dd953e5592c21e58be30d807
        depositAddressId:
          type: string
          description: 입금 주소 ID
          example: ae40b1b3dd953e5592c21e58be30d807
        ticker:
          type: string
          description: 암호화폐의 기호
          example: USDT
        transferType:
          enum:
            - WITHDRAWAL
            - DEPOSIT
          type: string
          description: 입출금 타입
          example: WITHDRAWAL
        transactionId:
          type: string
          description: >-
            트랜잭션 ID (Henesis Wallet에서 부여하는 트랜잭션의 고유 ID입니다. 온체인상 트랜잭션 해시와 다른
            개념입니다.)
          example: 8c87c578d7568edc156f831cf03c3ff0
        transactionHash:
          type: string
          description: 트랜잭션 해시
          example: '0xef76a243fa224f723922a1b067dd916fb1b2568aff292d2d1d183a807804922f'
        hopTransactionId:
          type: string
          description: >-
            홉 트랜잭션 ID (Henesis Wallet에서 부여하는 트랜잭션의 고유 ID입니다. 온체인상 트랜잭션 해시와 다른
            개념입니다.)
          example: 6f831cf03c3ff08c87c578d7568edc15
        hopTransactionHash:
          type: string
          description: 홉 트랜잭션 해시
          example: '0x60603c815f0ba0ad6b7f2ae398bd7cb9fe71347f984e32453fe4fe53f255dfd3'
        createdAt:
          type: string
          description: '트랜잭션 생성 시간 (형식: ms, UNIX time)'
          example: '1612411568760'
        updatedAt:
          type: string
          description: '트랜잭션 상태가 마지막으로 변경된 시간 (형식: ms, UNIX time)'
          example: '1612411724023'
        name:
          type: string
          description: 해당 내역의 지갑 또는 입금 주소의 이름
          example: ETH 실비 정산
        metadata:
          type: string
          description: 기타 정보 기록용 메타 데이터 (255자 제한)
          example: metadata
      required:
        - id
        - from
        - to
        - amount
        - status
        - orgId
        - decimals
        - walletId
        - depositAddressId
        - ticker
        - transferType
        - transactionId
        - transactionHash
        - hopTransactionId
        - hopTransactionHash
        - createdAt
        - updatedAt
        - name
    GetManyResponseForTransferDTO:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/TransferDTO'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
      required:
        - results
...
    NftDTO:
      type: object
      properties:
        id:
          type: number
          description: NFT 컨트랙트의 id
          example: 11
        name:
          type: string
          description: NFT 컨트랙트의 이름
          example: Dave Starbelly
        symbol:
          type: string
          description: NFT 컨트랙트의 심볼
          example: Dave Starbelly
        address:
          type: string
          description: NFT 컨트랙트의 주소
          example: '0xdd00383e5a51166bba3e96d84c14a36d72e6c128'
      required:
        - id
        - name
        - symbol
        - address
...
    NftTokenDTO:
      type: object
      properties:
        name:
          type: string
          description: NFT 토큰의 이름
          example: Haechi 1st NFT
        onchainId:
          type: string
          description: NFT 토큰의 온체인 ID
          example: '1'
        uri:
          type: string
          description: NFT 토큰의 URI
          example: https://opensea-creatures-api.herokuapp.com/api/creature/3
        externalUrl:
          type: string
          description: NFT 토큰 메타데이터 중 하나인 외부 주소
          example: https://openseacreatures.io/3
        imageUrl:
          type: string
          description: NFT 토큰의 이미지 주소
          example: https://storage.googleapis.com/opensea-prod.appspot.com/puffs/3.png
        metadata:
          type: object
          description: NFT 토큰의 메타데이터
          example: &ref_3
            description: Friendly OpenSea Creature that enjoys long swims in the ocean.
            external_url: https://openseacreatures.io/3
            image: >-
              https://storage.googleapis.com/opensea-prod.appspot.com/puffs/3.png
            name: Dave Starbelly
            attributes:
              - trait_type: Base
                value: Starfish
              - trait_type: Eyes
                value: Big
      required:
        - name
        - onchainId
        - uri
        - externalUrl
        - imageUrl
        - metadata
    NftTransferDTO:
      type: object
      properties:
        id:
          type: number
          description: NFT 입출금 내역의 ID
          example: 375031
        from:
          type: string
          description: 출금 주소
          example: '0x1f10ecbd971eab345ea19e96dc237b1fbd63de96'
        to:
          type: string
          description: 입금 주소
          example: '0xdd00383e5a51166bba3e96d84c14a36d72e6c128'
        status:
          enum:
            - PENDING_APPROVAL
            - REJECTED
            - REQUESTED
            - PENDING
            - FAILED
            - REVERTED
            - REPLACED
            - MINED
            - CONFIRMED
          type: string
          description: 트랜잭션 상태
          example: CONFIRMED
        orgId:
          type: string
          description: 지갑이 속한 팀(Organization)의 ID
          example: 31cafc79bd437e1b8e7b8209e399d3f3
        walletId:
          type: string
          description: 입출금 지갑 ID
          example: ae40b1b3dd953e5592c21e58be30d807
        depositAddressId:
          type: string
          description: 입금 주소 ID
          example: ae40b1b3dd953e5592c21e58be30d807
        nft:
          description: NFT 정보
          example: *ref_2
          allOf:
            - $ref: '#/components/schemas/NftDTO'
        token:
          description: NFT 토큰 정보
          example: &ref_27
            name: Haechi 1st NFT
            onchainId: '1'
            uri: https://opensea-creatures-api.herokuapp.com/api/creature/3
            externalUrl: https://openseacreatures.io/3
            imageUrl: >-
              https://storage.googleapis.com/opensea-prod.appspot.com/puffs/3.png
            metadata: *ref_3
          allOf:
            - $ref: '#/components/schemas/NftTokenDTO'
        transferType:
          enum:
            - WITHDRAWAL
            - DEPOSIT
          type: string
          description: 입출금 타입
          example: WITHDRAWAL
        transactionId:
          type: string
          description: >-
            트랜잭션 ID (Henesis Wallet에서 부여하는 트랜잭션의 고유 ID입니다. 온체인상 트랜잭션 해시와 다른
            개념입니다.)
          example: 8c87c578d7568edc156f831cf03c3ff0
        transactionHash:
          type: string
          description: 트랜잭션 해시
          example: '0xef76a243fa224f723922a1b067dd916fb1b2568aff292d2d1d183a807804922f'
        createdAt:
          type: string
          description: '트랜잭션 생성 시간 (형식: ms, UNIX time)'
          example: '1612411568760'
        updatedAt:
          type: string
          description: '트랜잭션 상태가 마지막으로 변경된 시간 (형식: ms, UNIX time)'
          example: '1612411724023'
        name:
          type: string
          description: 해당 내역의 지갑 또는 입금 주소의 이름
          example: ETH 실비 정산
        metadata:
          type: string
          description: 기타 정보 기록용 메타 데이터 (255자 제한)
          example: metadata
      required:
        - id
        - from
        - to
        - status
        - orgId
        - walletId
        - depositAddressId
        - nft
        - token
        - transferType
        - transactionId
        - transactionHash
        - createdAt
        - updatedAt
        - name
    GetManyResponseForNftTransferDTO:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/NftTransferDTO'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
      required:
        - results
...
paths:
...
  /api/v3/klaytn/transfers:
    get:
      operationId: klaytn-TransfersController_getTransfers
      x-readme:
        explorer-enabled: false
        samples-languages:
          - curl
          - node
          - java
          - python
          - go
      summary: 전체 입출금 목록 조회하기
      description: 모든 지갑의 가상자산 입출금 내역을 조회합니다.
      parameters:
        - name: X-Henesis-Secret
          in: header
          description: 대시보드를 통해 발급 받은 API Secret을 입력합니다.
          schema:
            type: string
        - name: Authorization
          in: header
          description: 대시보드를 통해 발급 받은 Access Token을 입력합니다. e.g. Bearer <TOKEN\>
          schema:
            example: Bearer <TOKEN>
            default: 'Bearer '
            type: string
        - name: ticker
          required: false
          in: query
          description: 코인의 기호
          schema:
            type: string
        - name: depositAddressId
          required: true
          in: query
          schema:
            type: string
        - name: walletId
          required: true
          in: query
          schema:
            type: string
        - name: userWalletId
          required: false
          in: query
          description: 사용자 지갑 ID
          schema:
            type: string
        - name: masterWalletId
          required: false
          in: query
          description: 마스터 지갑 ID
          schema:
            type: string
        - name: transactionId
          required: false
          in: query
          description: >-
            트랜잭션 ID (Henesis Wallet에서 부여하는 트랜잭션의 고유 ID입니다. 온체인상 트랜잭션 해시와 다른
            개념입니다.)
          schema:
            type: string
        - name: transactionHash
          required: false
          in: query
          description: 트랜잭션 해시
          schema:
            type: string
        - name: status
          required: false
          in: query
          description: 트랜잭션 상태
          schema:
            type: string
        - name: transferType
          required: false
          in: query
          description: 트랜잭션의 종류(입금, 출금)
          schema:
            type: string
        - name: updatedAtGte
          required: false
          in: query
          description: 'updatedAt이 해당 시점과 같거나 그보다 이후인 트랜잭션 조회 (형식: ms, UNIX time) '
          schema:
            type: string
        - name: updatedAtLt
          required: false
          in: query
          description: 'updatedAt이 해당 시점보다 이전인 트랜잭션 조회 (형식: ms, UNIX time) '
          schema:
            type: string
        - name: size
          required: false
          in: query
          description: '한 번의 요청으로 반환되는 응답의 개수 (기본값: 15)'
          schema:
            type: number
        - name: page
          required: false
          in: query
          description: 몇 번째 페이지를 조회할 것인지? (페이지는 0부터 시작)
          schema:
            type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetManyResponseForTransferDTO'
              example:
                pagination: *ref_11
                results:
                  - id: 375031
                    from: '0x1f10ecbd971eab345ea19e96dc237b1fbd63de96'
                    to: '0xdd00383e5a51166bba3e96d84c14a36d72e6c128'
                    amount: '1000000000'
                    status: CONFIRMED
                    orgId: 31cafc79bd437e1b8e7b8209e399d3f3
                    decimals: 18
                    walletId: ae40b1b3dd953e5592c21e58be30d807
                    depositAddressId: ae40b1b3dd953e5592c21e58be30d807
                    ticker: USDT
                    transferType: WITHDRAWAL
                    transactionId: 8c87c578d7568edc156f831cf03c3ff0
                    transactionHash: >-
                      0xef76a243fa224f723922a1b067dd916fb1b2568aff292d2d1d183a807804922f
                    hopTransactionId: 6f831cf03c3ff08c87c578d7568edc15
                    hopTransactionHash: >-
                      0x60603c815f0ba0ad6b7f2ae398bd7cb9fe71347f984e32453fe4fe53f255dfd3
                    createdAt: '1612411568760'
                    updatedAt: '1612411724023'
                    name: ETH 실비 정산
                    metadata: metadata
          description: ''
        '400':
          description: 올바르지 않은 트랜잭션 상태(status)로 요청하면 발생합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidStatusException'
              example: *ref_12
        '401':
          description: 인증이 실패했을 때 발생하는 에러입니다.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/InvalidAccessTokenException'
                  - $ref: '#/components/schemas/AccessTokenNotProvidedException'
                  - $ref: '#/components/schemas/InvalidAccessIpException'
              examples:
                InvalidAccessTokenException:
                  value: *ref_8
                AccessTokenNotProvidedException:
                  value: *ref_9
                InvalidAccessIpException:
                  value: *ref_10
      tags:
        - transfers
...
  /api/v3/klaytn/wallets/{masterWalletId}/nft/transfers:
    get:
      operationId: klaytn-WalletsController_getNftTransfers
      x-readme:
        explorer-enabled: false
        samples-languages:
          - curl
          - node
          - java
          - python
          - go
      summary: 전체 NFT 입출금 목록 조회하기
      description: 특정 지갑의 NFT 입출금 내역을 조회합니다.
      parameters:
        - *ref_17
        - *ref_18
        - name: masterWalletId
          required: true
          in: query
          description: 마스터 지갑 ID
          schema:
            type: string
        - name: nftId
          required: false
          in: query
          description: NFT 컨트랙트의 ID
          schema:
            type: number
        - name: tokenName
          required: false
          in: query
          description: NFT 토큰의 이름
          schema:
            type: string
        - name: tokenOnchainId
          required: false
          in: query
          description: NFT 토큰의 온체인 ID
          schema:
            type: string
        - name: userWalletId
          required: false
          in: query
          description: 사용자 지갑 ID
          schema:
            type: string
        - name: transactionId
          required: false
          in: query
          description: >-
            트랜잭션 ID (Henesis Wallet에서 부여하는 트랜잭션의 고유 ID입니다. 온체인상 트랜잭션 해시와 다른
            개념입니다.)
          schema:
            type: string
        - name: transactionHash
          required: false
          in: query
          description: 트랜잭션 해시
          schema:
            type: string
        - name: status
          required: false
          in: query
          description: 트랜잭션 상태
          schema:
            type: string
        - name: transferType
          required: false
          in: query
          description: 트랜잭션의 종류(입금, 출금)
          schema:
            type: string
        - name: updatedAtGte
          required: false
          in: query
          description: 'updatedAt이 해당 시점과 같거나 그보다 이후인 트랜잭션 조회 (형식: ms, UNIX time) '
          schema:
            type: string
        - name: updatedAtLt
          required: false
          in: query
          description: 'updatedAt이 해당 시점보다 이전인 트랜잭션 조회 (형식: ms, UNIX time) '
          schema:
            type: string
        - name: size
          required: false
          in: query
          description: '한 번의 요청으로 반환되는 응답의 개수 (기본값: 15)'
          schema:
            type: number
        - name: page
          required: false
          in: query
          description: 몇 번째 페이지를 조회할 것인지? (페이지는 0부터 시작)
          schema:
            type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetManyResponseForNftTransferDTO'
              example:
                pagination: *ref_11
                results:
                  - id: 375031
                    from: '0x1f10ecbd971eab345ea19e96dc237b1fbd63de96'
                    to: '0xdd00383e5a51166bba3e96d84c14a36d72e6c128'
                    status: CONFIRMED
                    orgId: 31cafc79bd437e1b8e7b8209e399d3f3
                    walletId: ae40b1b3dd953e5592c21e58be30d807
                    depositAddressId: ae40b1b3dd953e5592c21e58be30d807
                    nft: *ref_2
                    token: *ref_27
                    transferType: WITHDRAWAL
                    transactionId: 8c87c578d7568edc156f831cf03c3ff0
                    transactionHash: >-
                      0xef76a243fa224f723922a1b067dd916fb1b2568aff292d2d1d183a807804922f
                    createdAt: '1612411568760'
                    updatedAt: '1612411724023'
                    name: KLAY 실비 정산
                    metadata: metadata
          description: ''
        '400':
          description: 올바르지 않은 트랜잭션 상태(status)로 요청하면 발생합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidStatusException'
              example: *ref_12
        '401': *ref_20
      tags:
        - wallets
...